requirements.txt
fastapi==0.104.1
fastapi-cache2==0.2.1
pydantic==2.5.1
uvicorn==0.24.0
httpx==0.25.1
pandas==1.3.5
xgboost==2.0.2
scikit-learn==1.3.0
